requirements.txt
sentence-transformers==5.1.2
faiss-cpu==1.13.0
openai==2.24.0
